@@ -11,7 +11,7 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_hooks: ^0.21.2
-  lucide_flutter: ^0.544.0
+  lucide_flutter: ^0.545.0
 
 dev_dependencies:
   flutter_test:
