@@ -14,7 +14,7 @@ dependencies:
     path: packages/cliq_terminal
   cliq_ui:
     path: packages/cliq_ui
-  lucide_flutter: ^0.544.0
+  lucide_flutter: ^0.545.0
   go_router: ^16.0.0
   flutter_hooks: ^0.21.2
   shared_preferences: ^2.5.3
